// be/src/enums/api-status.ts
export enum ApiStatus {
    OK = 'OK',
    ERROR = 'ERROR',
    EMPTY_NAME = 'EMPTY NAME',
    EMPTY_SURNAME = 'EMPTY SURNAME',
    EMPTY_FULLNAME = 'EMPTY FULLNAME',
    EMPTY_LOGIN = 'EMPTY_LOGIN',
    EMPTY_PASSWORD = 'EMPTY_PASSWORD',
    EMPTY_PHONE_NUMBER = 'EMPTY_PHONE_NUMBER',
    EMPTY_COUNTRY_CODE = 'EMPTY_COUNTRY_CODE',
    EMPTY_CURRENCY_CODE = 'EMPTY_CURRENCY_CODE',
    EMPTY_LANGUAGE_CODE = 'EMPTY_LANGUAGE_CODE',
}